server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psutil==5.9.4
requests==2.28.1
GPUtil==1.4.0
python-dotenv==0.21.0
